testing_mcp/requirements.txt
anthropic
google-genai
openai
